+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This change log adheres to standards from [Keep a CHANGELOG](https://keepachange
 
 ## Unreleased
 
+### Added
+* [`no-string-refs`]: allow this.refs in > 18.3.0 ([#3807][] @henryqdineen)
+
+[#3807]: https://github.com/jsx-eslint/eslint-plugin-react/pull/3807
+
 ## [7.35.1] - 2024.09.02
 
 ### Fixed